# Dict: rehesh when there are too many tombstones