--- a/base/dict.jl
+++ b/base/dict.jl
@@ -351,7 +351,7 @@ ht_keyindex2!(h::Dict, key) = ht_keyindex2_shorthash!(h, key)[1]
 
     sz = length(h.keys)
     # Rehash now if necessary
-    if (h.count*3 + h.ndel) > sz*2
+    if (h.count + h.ndel)*3 > sz*2
         # > 2/3 full (including tombstones)
         rehash!(h, h.count > 64000 ? h.count*2 : h.count*4)
     end