@KristofferC 2023-03-18
`runbenchmarks("collection", vs = ":master")`