@nanosoldier 2023-03-18
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49045#issuecomment-1474931432) has completed - no performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/4e3a4dd_vs_0a9abc1/report.md).