@oscardssmith 2023-03-19
Makes sense to me! One thing I'd be interested in is whether 2/3rds or 3/4ths is the better cutoff, but other than that, this makes sense.