@petvana 2023-03-19
Just re-run for 3/4 ratio and it seems to be slightly slower on average
```
n=1000  17.565 μs (9 allocations: 8.14 KiB)
n=2000  35.468 μs (12 allocations: 16.95 KiB)
n=5000  112.744 μs (15 allocations: 51.39 KiB)
n=10000  318.783 μs (20 allocations: 119.67 KiB)
n=20000  681.511 μs (20 allocations: 119.67 KiB)
n=50000  1.799 ms (31 allocations: 528.16 KiB)
n=100000  3.749 ms (37 allocations: 1.05 MiB)
```